requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-genai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
click>=8.1.0
jinja2>=3.1.0
rich>=13.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
stripe>=5.0.0
Authlib>=1.3.0
httpx>=0.24.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
